refactor find or create
